update to new stable release v1.13
This is more or less a no-op as we already carried the meaningful
patches from this release as debian/patches.

Good thing is, now we've zero downstream patches.

We mostly adapt this release to not get told by user that we miss it
and it's important bug fixes ;)

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>
(cherry picked from commit 0e9eb3950b3bbdd9386a74c7ebbac9649aa4c83a)
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>